pyproject.toml
[project]
name = "genesis-v3"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite",
    "aiohttp",
    "apscheduler>=3.11,<4",
    "python-telegram-bot>=21.0",
    "litellm==1.78.7",  # pinned — 1.82.7/1.82.8 contain malware, do not upgrade without review
    "qdrant-client",
    "pyyaml",
    "httpx",
    "diskcache>=5.6",
    "beautifulsoup4",
    "html2text",
    "lxml",
    "flask>=3.0",
    "fastmcp>=2.0,<3.0",  # 3.x requires mcp>=1.24 with breaking API changes
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio",
    "aioresponses",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear (catches common gotchas)
    "SIM",  # flake8-simplify
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line length (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
# AZ plugin extensions use LoopData() as default arg — AZ convention, not ours to change
"az_plugins/**" = ["B008"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
